Add capability to load font from stream
The texture_font_t struct can now contain an FT_Stream as a location,
allowing fonts to be loaded from a generic stream.  This does have the
drawback of making the texture-font header dependent upon the FreeType
API.

As an example, this enables use of SDL_RWops to stream compressed files
on Android.